+++ app/src/main/java/de/tomorrow/weather/data/models/WeatherDataDto.kt
@@ -3,9 +3,9 @@ package de.tomorrow.weather.data.models
 import com.google.gson.annotations.SerializedName
 
 data class WeatherDataDto(
-    @SerializedName("latitude") val latitude: Double,
-    @SerializedName("longitude") val longitude: Double,
-    @SerializedName("elevation") val elevation: Double,
+    @SerializedName("latitude") val latitude: Float,
+    @SerializedName("longitude") val longitude: Float,
+    @SerializedName("elevation") val elevation: Float,
     @SerializedName("timezone") val timezone: String,
     @SerializedName("current_weather") val currentWeather: CurrentWeatherDto
 )
