+++ app/src/main/java/de/tomorrow/weather/data/network/WeatherApi.kt
@@ -8,8 +8,8 @@ import retrofit2.http.Query
 interface WeatherApi {
     @GET("/v1/forecast")
     suspend fun getWeather(
-        @Query("latitude") latitude: Double,
-        @Query("longitude") longitude: Double,
+        @Query("latitude") latitude: Float,
+        @Query("longitude") longitude: Float,
         @Query("timezone") timezone: String,
         @Query("current_weather") currentWeather: Boolean = true,
     ): Response<WeatherDataDto>
