+++ app/src/main/java/de/tomorrow/weather/data/repositories/WeatherRepository.kt
@@ -4,5 +4,5 @@ import de.tomorrow.weather.domain.weather.models.WeatherData
 import kotlinx.coroutines.flow.Flow
 
 interface WeatherRepository {
-    fun getWeather(name: String, lat: Double, lon: Double): Flow<WeatherData>
+    fun getWeather(name: String, lat: Float, lon: Float): Flow<WeatherData>
 }
