+++ app/src/main/java/de/tomorrow/weather/domain/weather/models/Location.kt
@@ -2,6 +2,6 @@ package de.tomorrow.weather.domain.weather.models
 
 data class Location(
     val name: String,
-    val latitude: Double,
-    val longitude: Double
+    val latitude: Float,
+    val longitude: Float
 )
