+++ app/src/main/java/de/tomorrow/weather/domain/weather/models/WeatherData.kt
@@ -2,10 +2,10 @@ package de.tomorrow.weather.domain.weather.models
 
 data class WeatherData(
     val name: String,
-    val latitude: Double,
-    val longitude: Double,
-    val elevation: Double,
-    val temperature: Double,
-    val windSpeed: Double,
+    val latitude: Float,
+    val longitude: Float,
+    val elevation: Float,
+    val temperature: Float,
+    val windSpeed: Float,
     val windDirection: Int
 )
